Fixes 2 cases where some logs were lost
When the main process logged and exited immediatelly, the LogRecord
would be left in the queue and not processed.
When the children sent LogRecords, the main process could exit before
processing them.

The fix is to wait a little when the Handler is closed so as to give a
chance to flush the queue. If the children keep sending messages, they
will be lost anyway as the Handler should not keep the process alive.