Compatibility with ProcessPoolExecutor
According to the python cookbook
(https://docs.python.org/3.8/howto/logging-cookbook.html#using-concurrent-futures-processpoolexecutor)
multiprocessing.Manager().Queue() should be use with
ProcessPoolExecutor.
If not there is a change of deadlock
(https://docs.python.org/3.8/library/multiprocessing.html#pipes-and-queues).

This change should not break existing code as
multiprocessing.Manager().Queue() can be used as drop-in replacement for
the regular multiprocessing.Queue()